Move tests around too
